chore(Chart): Create an amcharts service mock and expose to the consumers.
