+++ terminus-ui/chart/src/chart.component.md
@@ -126,5 +126,9 @@ Some helpers are exposed to assist with testing. These are imported from `@termi
 | `getChartDebugElement` |
 | `getChartInstance`     |
 
+|   Chart service mock    |
+|-------------------------|
+| `AmChartsServiceMock`   |
+
 
 [test-helpers-src]: https://github.com/GetTerminus/terminus-ui/blob/release/terminus-ui/chart/testing/src/test-helpers.ts
