+++ terminus-ui/chart/testing/src/test-helpers.ts
@@ -24,3 +24,61 @@ export function getChartInstance(fixture: ComponentFixture<any>): TsChartCompone
   const debugElement = getChartDebugElement(fixture);
   return debugElement.componentInstance;
 }
+
+/**
+ * Expose an amcharts service mock for consumers to use.
+ */
+
+export class AmChartsServiceMock {
+  public get amCharts() {
+    return {
+      core: {
+        color: jest.fn(),
+        options: {},
+        create: jest.fn(() => ({
+          series: {
+            clear: jest.fn(),
+            push: jest.fn(() => ({
+              dataFields: {},
+              tooltip: { background: {} },
+            })),
+          },
+          responsive: { enabled: false },
+          colors: { list: [] },
+          xAxes: {
+            push: jest.fn(() => ({
+              dataFields: {},
+              renderer: {
+                grid: { template: {} },
+                labels: { template: {} },
+              },
+            })),
+          },
+          yAxes: {
+            push: jest.fn(() => ({
+              title: {},
+              numberFormatter: {},
+              tooltip: {},
+            })),
+          },
+          dispose: jest.fn(),
+        })),
+      },
+      charts: {
+        CategoryAxis: jest.fn(),
+        ValueAxis: jest.fn(),
+        XYCursor: jest.fn(() => ({})),
+        Legend: jest.fn(() => ({})),
+        LineSeries: jest.fn(() => ({})),
+        ColumnSeries: jest.fn(() => ({})),
+        XYChart: {},
+        PieChart: {},
+        RadarChart: {},
+        TreeMap: {},
+        SankeyDiagram: {},
+        ChordDiagram: {},
+      },
+      maps: { MapChart: {} },
+    };
+  }
+}
